requirements.txt
aiohttp==3.7.4.post0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
fastapi==0.68.1
h11==0.12.0
idna==3.2
multidict==5.1.0
newsapi-python==0.2.6
pydantic==1.8.2
python-dotenv==0.19.0
requests==2.26.0
starlette==0.14.2
typing-extensions==3.10.0.2
urllib3==1.26.6
uvicorn==0.15.0
yarl==1.6.3
